DR rework. Eternity Mode change: Hard DR cap on 75%.
Internal restructure and rework for DR to be more structured.

All modifiers on PLAYER final damage are now additive, unless otherwise stated.
 For example:
if you have an accessory that decreases your damage taken by 15% and a debuff that increases it by 20% (for example Smite), their combined total effect will be increasing your damage taken by 5%. (+0.2 - 0.15 = +0.05)

This new system is way more standardized. consistent and intuitive. A custom DR formula is easy to do if needed, but would prefer that as only a last resort as it's much less intuitive.